+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Last changes:
 
 
+Version `7 Nov'23`
+
+  - NEW: dat_clr_test.tcl (for apave 4.2.0a2+)
+
+
 Version `17 Dec'22`
 
   - NEW: apave's tests separated into apave_tests repo